+++ CHANGELOG.md
@@ -1,5 +1,7 @@
 # Changelog
 
+- rename gitignore, gitattributes and npmignore templates to fix missing template issue ([#1](https://github.com/seaofvoices/generator-luau/pull/1))
+
 ## 0.1.0
 
 - Initial version